improved support for qwen

+++ interpretability_backend/backend/embedding_functions/specific_functions/embed_qwen.py
@@ -25,16 +25,14 @@ class EmbedTextQWEN(EmbeddingFunction[Documents]):
         self.is_query = is_query
 
         self.tokeniser = AutoTokenizer.from_pretrained(model, padding_side="left")
-        ## If Cuda is available, use Flash attention, else avoid it
-        ## To note that I'm using fp32 for mps if we don't use flash attention
-        ## So it's possible that we should let for cuda also let the possibility of using fp32
-        ## But I should test how much the performance degrades with fp16/32
+        # CUDA enables Flash Attention 2; MPS/CPU use the default attention impl.
+        # bfloat16 matches Qwen3's training dtype (fp16 risks overflow on long-context pooling).
         self.model = (
             AutoModel.from_pretrained(
-                model, attn_implementation="flash_attention_2", torch_dtype=torch.float16
+                model, attn_implementation="flash_attention_2", torch_dtype=torch.bfloat16
             ).cuda()
             if device == "cuda"
-            else AutoModel.from_pretrained(model, torch_dtype=torch.float16).to(device)
+            else AutoModel.from_pretrained(model, torch_dtype=torch.bfloat16).to(device)
         )
         self.model.eval()
 
@@ -85,8 +83,10 @@ class EmbedTextQWEN(EmbeddingFunction[Documents]):
                 ## I leave the normalisation as an option
                 embeddings = F.normalize(embeddings, p=2, dim=1)
 
-        ### Embedidng type wants a list of np.ndarray
-        ### it returns a np array containing that contains other arrays
-        ### we explictely convert to np.array first for the type checker
+        ### Embedding type wants a list of np.ndarray.
+        ### .numpy() returns a single np.ndarray (rows = vectors); list(...) splits it
+        ### into the per-row arrays the type checker expects.
+        ### bfloat16 has no numpy equivalent, so we cast to float32 first
+        ### (ChromaDB's storage dtype, so no extra conversion downstream).
 
-        return list(embeddings.detach().cpu().numpy())
+        return list(embeddings.detach().to(torch.float32).cpu().numpy())
